requirements.txt
inky[rpi]
git+https://github.com/kipe/nordpool.git@0.3.3
matplotlib
numpy
pandas
Pillow
